+++ src/components/SmallSelect/styles.tsx
@@ -1,5 +1,6 @@
-import styled, { css, keyframes } from 'styled-components';
+import styled, { keyframes } from 'styled-components';
 import ReactSelect from 'react-select';
+import { SmallSelectProps } from '.';
 
 const animateCheckbox = keyframes`
   0% {
@@ -10,7 +11,7 @@ const animateCheckbox = keyframes`
   }
 `;
 
-export const Select = styled(ReactSelect)<ReactSelect>`
+export const Select = styled(ReactSelect)<SmallSelectProps>`
   .select__control {
     min-height: 2rem;
     max-height: 2rem;
@@ -18,37 +19,59 @@ export const Select = styled(ReactSelect)<ReactSelect>`
     margin: 0;
     box-sizing: border-box;
     border-radius: 0.75rem;
-    border-color: rgba(20, 20, 20, 0.7);
+    border-color: #5b5b5b;
     transition: none;
     background-color: transparent;
     &:hover {
-      border-color: rgba(20, 20, 20, 1);
+      border-color: #5b5b5b;
     }
+
+    .select__value-container.select__value-container--is-multi.select__value-container--has-value {
+      flex-wrap: wrap;
+      overflow: hidden;
+      max-height: 1.8rem;
+      margin-bottom: 0rem;
+    }
+
     &--menu-is-open {
+      max-height: 25rem;
       border-bottom-left-radius: 0;
       border-bottom-right-radius: 0;
       border-bottom: none;
       padding-bottom: 1px;
-      .select__dropdown-indicator {
-        transform: rotate(180deg);
+
+      .select__value-container.select__value-container--is-multi.select__value-container--has-value {
+        flex-wrap: wrap;
+        overflow: hidden;
+        max-height: 20rem;
       }
     }
     &--is-focused {
       box-shadow: none;
-      border-color: rgba(20, 20, 20, 1);
+      border-color: #141414;
+    }
+    &--menu-is-open .select__indicators svg {
+      transform: rotate(180deg);
     }
   }
   .select__value-container {
-    padding: 0 0.5rem;
+    padding: 0 0.5rem 0 0.3rem;
   }
   .select__indicators {
-    padding: 0 0.75rem 0 0;
+    position: relative;
+    align-items: unset;
+
+    > svg {
+      position: absolute;
+      top: 0.8125rem;
+      right: 0.625rem;
+    }
   }
   .select__dropdown-indicator {
-    color: rgba(20, 20, 20, 0.7);
+    color: #5b5b5b;
     padding: 0;
     &:hover {
-      color: rgba(20, 20, 20, 1);
+      color: #141414;
     }
   }
   .select__menu {
@@ -56,7 +79,7 @@ export const Select = styled(ReactSelect)<ReactSelect>`
     border: none;
     box-shadow: none;
     border-radius: 0;
-    border: 1px solid rgba(20, 20, 20, 0.7);
+    border: 1px solid #5b5b5b;
     border-top: none;
     border-radius: 0 0 0.75rem 0.75rem;
     padding: 0 0 0.25rem 0;
@@ -70,8 +93,8 @@ export const Select = styled(ReactSelect)<ReactSelect>`
     &::-webkit-scrollbar-track {
       border-radius: 0;
       /* border-bottom-right-radius: 0.75rem; */
-      background-color: rgba(20, 20, 20, 0.1);
-      box-shadow: 0 0.25rem 0 0 rgba(20, 20, 20, 0.1);
+      background-color: #e7e7e7;
+      box-shadow: 0 0.25rem 0 0 #e7e7e7;
     }
 
     &::-webkit-scrollbar {
@@ -83,7 +106,7 @@ export const Select = styled(ReactSelect)<ReactSelect>`
 
     &::-webkit-scrollbar-thumb {
       border-radius: 10px;
-      border: 2px solid rgb(231, 231, 231);
+      border: 2px solid #e7e7e7;
       box-sizing: border-box;
       background-color: #b3b2b1;
       margin: 2px;
@@ -91,20 +114,24 @@ export const Select = styled(ReactSelect)<ReactSelect>`
   }
   .select__placeholder {
     font-size: 0.6875rem;
-    color: rgba(20, 20, 20, 0.3);
+    color: #b9b9b9;
+    margin-left: 0.4375rem;
   }
   .select__single-value,
   .select__input {
     font-size: 0.6875rem;
-    color: rgba(20, 20, 20, 1);
-    font-weight: 500;
+    margin-left: 0.4375rem;
+    color: #141414;
     > input {
-      font-weight: 500;
+      font-weight: 700;
     }
+  }
+  .select__single-value-value {
+    font-weight: 500;
   }
   .select__option {
     font-size: 0.875rem;
-    color: rgba(20, 20, 20, 0.7);
+    color: #5b5b5b;
     padding: 0.5rem 1rem;
     font-weight: 400;
     display: flex;
@@ -119,17 +146,17 @@ export const Select = styled(ReactSelect)<ReactSelect>`
     }
     &--is-selected {
       font-weight: 500;
-      color: rgba(20, 20, 20, 1);
+      color: #141414;
       background-color: unset;
     }
     & + .select__option {
-      border-top: 1px solid rgba(20, 20, 20, 0.1);
+      border-top: 1px solid #e7e7e7;
     }
   }
   .select__multi-value {
     background-color: #cee9dd;
     border-radius: 0.25rem;
-    margin: 0 0.125rem;
+    margin: 0.125rem;
     padding: 0 0.5rem;
     align-items: center;
     height: 1.5rem;
@@ -160,4 +187,8 @@ export const Select = styled(ReactSelect)<ReactSelect>`
       }
     }
   }
+`;
+
+export const SelectWrapper = styled.div`
+  position: relative;
 `;