+++ src/components/Table/styles.ts
@@ -1,6 +1,6 @@
 import styled, { css, keyframes } from 'styled-components';
 import { StyledArrow } from '../Icons/Arrow/styles';
-import { ITable, ITableStyles, ITableColumnOptions, ITableHoverable, StyledHeaderProps } from './interfaces';
+import { ITable, ITableStyles, ITableColumnOptions, ITableHoverable } from './interfaces';
 
 export const tableDefaultStyles: ITable = {
   colors: {
@@ -61,6 +61,7 @@ export const Column = styled.td<ITableColumnOptions>`
 
 export const Row = styled.tr`
   position: relative;
+  display: table-row !important;
 `;
 
 const tableHeaderLoadingAnimation = keyframes`
@@ -81,6 +82,8 @@ const tableHeaderLoadingAnimation = keyframes`
 export const Head = styled.thead`
   background-color: ${({ theme }) =>
     theme?.table?.colors?.head?.backgroundColor || tableDefaultStyles?.colors?.head?.backgroundColor};
+
+  box-shadow: #f7f6f4 0px 0.75rem 0px;
 
   ${Column} {
     color: ${({ theme }) => theme?.table?.colors?.head?.color || tableDefaultStyles?.colors?.head?.color};
@@ -148,17 +151,15 @@ export const Body = styled.tbody<ITableStyles>`
       &:last-child:not(:first-child) {
         border-left: 0;
       }
-    }
 
-    &:last-of-type {
-      ${Column} {
-        &:first-child {
-          border-bottom-left-radius: 12px;
-        }
+      &:first-child {
+        border-bottom-left-radius: 12px;
+        border-top-left-radius: 12px;
+      }
 
-        &:last-child {
-          border-bottom-right-radius: 12px;
-        }
+      &:last-child {
+        border-bottom-right-radius: 12px;
+        border-top-right-radius: 12px;
       }
     }