+++ src/index.tsx
@@ -33,6 +33,12 @@ import TagSelect from './components/TagSelect';
 import InlineLoader from './components/InlineLoader';
 import Breadcrumbs from './components/Breadcrumbs';
 import Orderer from './components/Orderer';
+import Infotip from './components/Infotip';
+import DatePicker from './components/DatePicker';
+import FileUpload from './components/FileUpload';
+import ProgressBar from './components/ProgressBar';
+import Select from './components/Select';
+import LabelSwitch from './components/LabelSwitch';
 
 export {
   WizardSteps,
@@ -70,4 +76,10 @@ export {
   InlineLoader,
   Breadcrumbs,
   Orderer,
+  Infotip,
+  DatePicker,
+  FileUpload,
+  ProgressBar,
+  Select,
+  LabelSwitch,
 };